requirements.txt
Django
Pillow
djangorestframework
markdown
django-extensions
django-filter
django-cors-headers